[CARBONDATA-1856][PARTITION] Support insert/load data for partition table
Changed carbonrelation to HadoopFSRelation during load in case of the partition table.
Implement sparks Fileformat interface for carbon and use carbonoutputformat inside.
Create partition.map file inside each segment for mapping between partition and index file.

This closes #1654

+++ integration/spark-common/src/main/scala/org/apache/carbondata/spark/rdd/UpdateDataLoad.scala
@@ -35,7 +35,7 @@ object UpdateDataLoad {
 
   def DataLoadForUpdate(
       segId: String,
-      index: Int,
+      index: Long,
       iter: Iterator[Row],
       carbonLoadModel: CarbonLoadModel,
       loadMetadataDetails: LoadMetadataDetails): Unit = {
